Readme file Issue Updation
### Updated README Structure:

Added Navigation: A new navigation section was included at the top of the README to enhance user experience. This allows users to quickly navigate to important sections such as Project Overview, Tech Stack, How to Fork, How to Contribute, Code of Conduct, and Contact Information.
Revised Contribution Guide: Clearer instructions were added for contributing to the project, including step-by-step details on how to fork the repository, create new branches, make changes, and submit a pull request.
Included Code of Conduct: A section for the project's Code of Conduct was added, ensuring contributors are aware of community guidelines and acceptable behavior.
Contact Information: A contact section was introduced for any inquiries or support, directing users where they can reach out for help.
Other Visual Enhancements: The layout and formatting were improved for better readability and aesthetics, including the addition of icons, badges, and emojis to make the README visually appealing.

### Please label this under GSSOC'24 extd and Hacktoberfest-accepted and pls add the label of the level I did contribute.